[JENKINS-57800] Add Jenkins parent pom and other pom fixes
