--- plugin-management-cli/src/main/java/library/Main.java
@@ -1,14 +1,0 @@
-package library;
-
-import pluginmanager.PluginManager;
-
-
-public class Main {
-    public static void main(String[] args) {
-
-        PluginManager pm = new PluginManager();
-        pm.start();
-
-    }
-
-}
